+++ database/migrations/2025_12_30_130004_add_discount_to_products_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->decimal('discount', 20, 2)->default(0);
+            $table->integer('discount')->default(0);
         });
     }
